# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRM Doc API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tasks": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Index",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "title example"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "description example"
                                                    },
                                                    "priority": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "active"
                                                    },
                                                    "client_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "project_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "title example"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "description example"
                                            },
                                            "priority": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "active"
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "client_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "project_id": {
                                                "type": "integer",
                                                "example": 2
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "title example"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "description example"
                                                },
                                                "priority": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "active"
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "client_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "project_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tasks/{task}": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Show",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "Task's id",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "title example"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "description example"
                                                    },
                                                    "priority": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "active"
                                                    },
                                                    "client_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "project_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Task"
                ],
                "summary": "Delete",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "Task's id",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Task"
                ],
                "summary": "Update",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "Task's id",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "title example"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "description example"
                                            },
                                            "priority": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "active"
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "client_id": {
                                                "type": "integer",
                                                "example": 2
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "title example"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "description example"
                                                },
                                                "priority": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "active"
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "client_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}